requirements.txt
3==0.0.0
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0.2
django-elasticsearch-dsl==7.2.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
docopt==0.6.2
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
Pillow==8.4.0
pipreqs==0.4.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
suds==1.0.0

txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
yarg==0.1.9
zope.interface==5.4.0
